requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python
aiofiles
rembg
python-multipart